# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * China vs USA Asset & Currency Comparison API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ExchangeRate {
  usdToCny: number;
  cnyToUsd: number;
  change24h: number;
  changePercent24h: number;
  high24h: number;
  low24h: number;
  lastUpdated: string;
}

export interface AssetItem {
  name: string;
  value: number;
  change: number;
  changePercent: number;
  unit: string;
  /** @nullable */
  description?: string | null;
}

export interface UsaAssets {
  sp500: AssetItem;
  usdStrength: AssetItem;
  techMarketCap: AssetItem;
  gdp: AssetItem;
  inflation: AssetItem;
  unemployment: AssetItem;
}

export interface ChinaAssets {
  shanghaiIndex: AssetItem;
  cnyStrength: AssetItem;
  manufacturingOutput: AssetItem;
  gdp: AssetItem;
  inflation: AssetItem;
  tradeBalance: AssetItem;
}

export type ComparisonRowWinner = typeof ComparisonRowWinner[keyof typeof ComparisonRowWinner];


export const ComparisonRowWinner = {
  USA: 'USA',
  CHINA: 'CHINA',
  TIE: 'TIE',
} as const;

export interface ComparisonRow {
  metric: string;
  usa: string;
  china: string;
  usaValue: number;
  chinaValue: number;
  winner: ComparisonRowWinner;
  unit: string;
  /** @nullable */
  description?: string | null;
}

export type ComparisonDataOverallWinner = typeof ComparisonDataOverallWinner[keyof typeof ComparisonDataOverallWinner];


export const ComparisonDataOverallWinner = {
  USA: 'USA',
  CHINA: 'CHINA',
  TIE: 'TIE',
} as const;

export interface ComparisonData {
  rows: ComparisonRow[];
  overallWinner: ComparisonDataOverallWinner;
  usaScore: number;
  chinaScore: number;
  lastUpdated: string;
}

export interface ChartPoint {
  timestamp: string;
  value: number;
}

export interface ChartData {
  period: string;
  points: ChartPoint[];
  change: number;
  changePercent: number;
}

export interface MarketsChartPoint {
  timestamp: string;
  usa: number;
  china: number;
}

export interface MarketsChartData {
  period: string;
  points: MarketsChartPoint[];
}

export type WinnerIndicatorWinner = typeof WinnerIndicatorWinner[keyof typeof WinnerIndicatorWinner];


export const WinnerIndicatorWinner = {
  USA: 'USA',
  CHINA: 'CHINA',
  TIE: 'TIE',
} as const;

export interface WinnerIndicator {
  category: string;
  winner: WinnerIndicatorWinner;
  margin: string;
  summary: string;
}

export type DashboardSummaryTodaysWinner = typeof DashboardSummaryTodaysWinner[keyof typeof DashboardSummaryTodaysWinner];


export const DashboardSummaryTodaysWinner = {
  USA: 'USA',
  CHINA: 'CHINA',
  TIE: 'TIE',
} as const;

export interface DashboardSummary {
  todaysWinner: DashboardSummaryTodaysWinner;
  aiInsight: string;
  categories: WinnerIndicator[];
  exchangeRate: number;
  usaStrength: number;
  chinaStrength: number;
  lastUpdated: string;
}

export type NewsArticleCountry = typeof NewsArticleCountry[keyof typeof NewsArticleCountry];


export const NewsArticleCountry = {
  USA: 'USA',
  CHINA: 'CHINA',
} as const;

export type NewsArticleCategory = typeof NewsArticleCategory[keyof typeof NewsArticleCategory];


export const NewsArticleCategory = {
  Markets: 'Markets',
  Economy: 'Economy',
  Currency: 'Currency',
  Trade: 'Trade',
  Technology: 'Technology',
  Policy: 'Policy',
} as const;

export type NewsArticleSentiment = typeof NewsArticleSentiment[keyof typeof NewsArticleSentiment];


export const NewsArticleSentiment = {
  positive: 'positive',
  negative: 'negative',
  neutral: 'neutral',
} as const;

export interface NewsArticle {
  id: string;
  title: string;
  summary: string;
  source: string;
  country: NewsArticleCountry;
  category: NewsArticleCategory;
  publishedAt: string;
  sentiment: NewsArticleSentiment;
  url: string;
  /** @nullable */
  imageTag?: string | null;
}

export interface NewsResponse {
  articles: NewsArticle[];
  lastUpdated: string;
  usaCount: number;
  chinaCount: number;
}

export interface ConversionResult {
  from: string;
  to: string;
  amount: number;
  result: number;
  rate: number;
}

export type GetCurrencyChartParams = {
/**
 * Time period for chart data
 */
period?: GetCurrencyChartPeriod;
};

export type GetCurrencyChartPeriod = typeof GetCurrencyChartPeriod[keyof typeof GetCurrencyChartPeriod];


export const GetCurrencyChartPeriod = {
  '1D': '1D',
  '7D': '7D',
  '30D': '30D',
} as const;

export type GetMarketsChartParams = {
/**
 * Time period for chart data
 */
period?: GetMarketsChartPeriod;
};

export type GetMarketsChartPeriod = typeof GetMarketsChartPeriod[keyof typeof GetMarketsChartPeriod];


export const GetMarketsChartPeriod = {
  '1D': '1D',
  '7D': '7D',
  '30D': '30D',
} as const;

export type GetNewsParams = {
/**
 * Filter by country
 */
country?: GetNewsCountry;
};

export type GetNewsCountry = typeof GetNewsCountry[keyof typeof GetNewsCountry];


export const GetNewsCountry = {
  USA: 'USA',
  CHINA: 'CHINA',
  ALL: 'ALL',
} as const;

export type ConvertCurrencyParams = {
/**
 * Amount to convert
 */
amount: number;
/**
 * Source currency
 */
from: ConvertCurrencyFrom;
};

export type ConvertCurrencyFrom = typeof ConvertCurrencyFrom[keyof typeof ConvertCurrencyFrom];


export const ConvertCurrencyFrom = {
  USD: 'USD',
  CNY: 'CNY',
} as const;