# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: China vs USA Asset & Currency Comparison API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /charts/currency:
    get:
      operationId: getCurrencyChart
      tags: [charts]
      summary: Get USD/CNY historical chart data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [1D, 7D, 30D]
          description: Time period for chart data
      responses:
        "200":
          description: Currency chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartData"
...
  /currency/convert:
    get:
      operationId: convertCurrency
      tags: [currency]
      summary: Convert between USD and CNY
      parameters:
        - name: amount
          in: query
          required: true
          schema:
            type: number
          description: Amount to convert
        - name: from
          in: query
          required: true
          schema:
            type: string
            enum: [USD, CNY]
          description: Source currency
      responses:
        "200":
          description: Conversion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversionResult"
components:
  schemas:
...
    ChartPoint:
      type: object
      required:
        - timestamp
        - value
      properties:
        timestamp:
          type: string
        value:
          type: number
    ChartData:
      type: object
      required:
        - period
        - points
        - change
        - changePercent
      properties:
        period:
          type: string
        points:
          type: array
          items:
            $ref: "#/components/schemas/ChartPoint"
        change:
          type: number
        changePercent:
          type: number
...
    ConversionResult:
      type: object
      required:
        - from
        - to
        - amount
        - result
        - rate
      properties:
        from:
          type: string
        to:
          type: string
        amount:
          type: number
        result:
          type: number
        rate:
          type: number