# lib/api-zod/src/generated/types/conversionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * China vs USA Asset & Currency Comparison API
 * OpenAPI spec version: 0.1.0
 */

export interface ConversionResult {
  from: string;
  to: string;
  amount: number;
  result: number;
  rate: number;
}